* PCとモバイルで表示を分離 * モバイルでは検索機能をいったん除去。あとでハンバーガーメニューとかで復活させたい
