+++ src/components/Common/Common.css
@@ -1,0 +1,63 @@
+﻿div.icon {
+  width: 23px;
+  height: 23px;
+}
+
+i.urlicon {
+  padding: 2px 2px 2px 2px;
+  vertical-align: middle;
+  margin: 0 auto;
+}
+
+i.urlicon.home {
+  color: #888888;
+}
+
+i.urlicon.steam {
+  color: #324b6c;
+}
+
+i.urlicon.twitter {
+  color: #80cafd;
+}
+
+i.urlicon.facebook {
+  color: #4267b2;
+}
+
+i.urlicon.youtube {
+  color: #ff0000;
+}
+
+img.urlicon {
+  width: 17px;
+  height: 17px;
+  padding: 0;
+  vertical-align: middle;
+  margin: 0 auto;
+  border-radius: 50%;
+  border-width: 1px;
+  border-style: solid;
+  border-color: #dddddd;
+}
+
+span.genre {
+  font-size: 0.6rem;
+  border-radius: 5px 5px 5px 5px;
+  border-style: none;
+  background-color: #fbcb2d;
+  color: #000000;
+  margin: 1px 1px 0px 0px;
+  padding: 1px 5px 0px 5px;
+  display: inline-block;
+}
+
+img.actor {
+  border-radius: 20%;
+  border-style: solid;
+  border-width: 1px;
+  border-color: #dddddd;
+  width: 15px;
+  height: 15px;
+  margin: 0px 2px 0px 0px;
+}
